+++ CHANGELOG.md
@@ -18,11 +18,8 @@ We use the following categories for changes:
 
 ### Added
 
-- added watch pvc readiness [#sc-4860]
-- validate node schedulable [#sc-4878]
-
 ### Changed
-- change timeout for portal connection polling to 7 mins [#sc-4860]
+
 ### Fixed
 
 ### Removed
@@ -30,6 +27,17 @@ We use the following categories for changes:
 ### Deprecated
 
 ### Security
+
+## [0.3.10] 2022-09-21
+
+### Added
+
+- added watch pvc readiness [#sc-4860]
+- validate node schedulable [#sc-4878]
+
+### Changed
+
+- change timeout for portal connection polling to 7 mins [#sc-4860]
 
 ## [0.3.9] 2022-09-20